==> jupyter_notebooks/01 - Data Collection.ipynb <==
# **Data Collection**

## Objectives

* Collect the dataset from kaggle and save as raw data
* Inspect the dataset and save it under output/datasets/collection

## Inputs

* Kaggle JSON file - the authentication token.

## Outputs

* Generate Dataset: outputs/datasets/collection/StudentPreformance.csv

## Additional Comments



---

# Change working directory

* We are assuming you will store the notebooks in a subfolder, therefore when running the notebook in the editor, you will need to change the working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Fetch data from Kaggle

Install Kaggle package to fetch data

%pip install kaggle==1.5.12

**Change kaggle configuration directory to current working directory and permission of kaggle authentication json.**

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

In the block belowe is the path for the Kaggle students preformens difined and a directory for it is ceatedasz, and then use a Kaggle command to download the dataset into this directory.

KaggleDatasetPath = "lainguyn123/student-performance-factors"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

Next step is to unzip the downloaded file and therafter delet the zip-fil and the kaggle.json file 

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

---

# Load and Inspect Kaggle data

The dataset is opened using the libary Pandas and then displays the first 5 rows in the dataset 

import pandas as pd
df = pd.read_csv(f"inputs/datasets/raw/StudentPerformanceFactors.csv")
df.head()

---

Belowe is a summary of the DataFrame

df.info()

Checks if the Dataframe have any *null* value

df.isnull().sum().sort_values(ascending=False)

new_list = df.duplicated(subset=None)

number_of_duplicted = 0
rows = []

for i in range(len(new_list)):
    if new_list[i]:
        number_of_duplicted += 1
        rows.append(i)

if number_of_duplicted == 0:
    print('There are no duplicates in the dataset')
else:
    print(f'{number_of_duplicted} rows are duplicated and the duplecations are on row {rows}')

## Conclusion
With a quick inspection for the dataset the following this was found 

* The dataset used has 20 variables where there is 1 target and 19 features. 
* None of the rows are duplicated 
* The dataset has missing data in the columns Parental_Education_Level, Teacher_Quality, Distance_from_Home with the corresponding rows missing data 90, 78, 67. 
    * Hopefully, these data rows are correlated and the reduction of the data set is as small as possible. However, the maximum loss can be 235 rows, which is a reduction of about 3%, which is a minimal reduction of the data set. 






---

# Push files to Repo

import os
try:
  # create outputs/datasets/collection folder
  os.makedirs(name='outputs/datasets/collection') # create outputs/datasets/collection folder
except Exception as e:
  print(e)

# Save the dataset in newlt created folder 
df.to_csv(f"outputs/datasets/collection/StudentPerformance.csv",index=False)
